1st sprint updated code of users
