+++ src/Admin/Users/UserAppbar.jsx
@@ -12,8 +12,8 @@ import {
   useTheme,
 } from "@mui/material";
 import { Search, Add } from "@mui/icons-material";
-import CustomModal from "../Admin/CommonComponents/CustomModal";
-import AddUser from "../Admin/Dashboard/AddUser";
+import CustomModal from "../CommonComponents/CustomModal";
+import AddUser from "../Dashboard/AddUser";
 
 const AppBarComponent = ({ searchQuery, setSearchQuery, userCount }) => {
   const [openUser, setOpenUser] = useState(false);
